requirements.txt
black
flake8
isort
loguru
mkdocs
pip
python-dotenv
tqdm
typer
pandas==2.0.0
numpy==1.24.3
polars
web3
hydra-core
statsmodels
pyarrow
pytest
ruff
-e .
dune-client
scikit-learn